refactor: component, tableRow 중복 제거
